Create 'Hello, World' skeleton app for Mass PFML (#1)
This creates a barebones Next.js app to get started on the Mass PFML Portal. The README includes prerequisites and run instructions. The packages.js file has been edited for relevance.

The boilerplate code is generated, with the exception of the pages/index.js React component and the script stanza in the packages.js file.

Co-authored-by: Sawyer Hollenshead <git@sawyerh.com>