Add js linting, code formatting, and pre-commit hook (#5)
This PR adds configuration for .js and .jsx linting using ESlint, as well as code formatting using Prettier. A git pre-commit hook is set up to run both formatting and linting. It also establishes a top-level package.json to configure common tools across the mono-repo, and adds documentation of these tools.

The ESLint rules have been copied from the Vermont project (https://github.com/navahq/archive-vermont-customer-portal-apps/blob/dev/.eslintrc) and should be modified as needed for the Mass project going forward.

I have omitted stylelint and CSS linting in general as we do not yet have a style system to lint against for this project. That can be added using the examples in the Vermont project (https://github.com/navahq/archive-vermont-customer-portal-apps/blob/dev/package.json) once it is available. I did set up linting for jest, jsdoc, and react-hooks in anticipation of their use in this project.

The index.js file was rewritten by Prettier and an import added as required by the linter.

Co-authored-by: Sawyer Hollenshead <git@sawyerh.com>