Add workflow for Portal continuous integration testing (#17)
## 🎫 Ticket

https://trello.com/c/0B8E2z9X/81-engineers-can-run-automated-tests

## 🛠 Changes

This adds a workflow that triggers testing on pull_requests. The output of the workflow is shown in several ways:
- In the Actions tab for the repository.
- Status is indicated by a green check, yellow circle, or red x to the right of the Pull Request title, indicating success, pending, or failure respectively.
- In the Checks tab on the pull request.

This check triggers on pull_request events, which includes `opened`, `synchronize`, and `reopened` activities. It only triggers on the `portal/` directory.

## ℹ️ Context for reviewers

This is our first step in automating testing. Once changes are pushed to the master branch, we can add a badge to the README showing test status.

Thanks to @sawyerh for getting this work off the ground.

## 👀 Demo

Here's what a failing check looks like on the PR list page:
<img width="625" alt="Screen Shot 2020-01-28 at 4 48 54 PM" src="https://user-images.githubusercontent.com/9326892/73318450-4a223880-41ee-11ea-889c-19b8fc71dbf8.png">

Here's what is shown on a PR with failing tests. Nice and loud to stop a dev from unknowingly submitting.
<img width="779" alt="Screen Shot 2020-01-28 at 4 53 37 PM" src="https://user-images.githubusercontent.com/9326892/73318624-da607d80-41ee-11ea-92c1-884c5f2bad61.png">