Initial CloudFront config (#23)
## 🎫 Ticket

https://trello.com/c/uWl8dxGd

## 🛠 Changes

Add Terraform code to create a CloudFront distribution for Portal S3 bucket. Unknown if the setup will be exactly like this for Mass account, e.g., possible/likely we won't have a hosted zone in Route53 to automatically point a nicer name to the CloudFront distribution.

## ℹ️ Context for reviewers

Had a CloudFront config sitting around from another project, just throwing it up as a starting place if it's helpful.