Add code coverage thresholds (#28)
## 🎫 [Ticket](https://trello.com/c/P7KgdB5P/84-engineers-can-see-test-coverage-of-automated-tests)

## 🛠 Changes

Configures Jest to have some minimum code coverage expectations. If these aren't met then CI will fail.

If for some reason we want to bypass coverage thresholds then alternative thresholds can be added for specific files/directories.

This only tests the overall coverage results, not per file. So if a new file is added and its coverage is 0% but that doesn't drag the average coverage down below the thresholds then the coverage will still be considered sufficient.

I explored adding a requirement where all .js files have to pass some minimum thresholds but it doesn't work well in combination with alternative thresholds for specific files (which are useful for bypassing the coverage standards deliberately). In order for the coverage to pass it would have to pass both the file-specific thresholds and then also the .js thresholds.

## ℹ️ Context for reviewers

Ideally we would integrate with some sort of code coverage tracking service so that it would be easy to compare a PR's coverage compared to that of master, and also track trends over time. It seems like Codecov, CodeClimate, and Coverall are popular choices. Because they cost money I've decided to stick with this for now and to bring up coverage tracking separately.

## 👀 Demo

Here's what it looks like when coverage fails:
<img width="717" alt="Screen Shot 2020-02-03 at 10 42 14 AM" src="https://user-images.githubusercontent.com/9039672/73670785-dcf02680-4677-11ea-964c-03c96c1097f2.png">

<img width="954" alt="Screen Shot 2020-02-03 at 10 41 45 AM" src="https://user-images.githubusercontent.com/9039672/73670800-e2e60780-4677-11ea-85a9-36ad524a3577.png">