CP-134: Add auth screen message integration with login screen messages (#80)
## 🎫 Ticket

[CP-134](https://lwd.atlassian.net/browse/CP-134)

## 🛠 Changes

This sets up the structure for auth screen messages to get customized. It also adds the customization for messages on the login screen specifically.

## ℹ️ Context for reviewers

EDIT: All i18n is now centralized in the `src/locales/i18n.js` module, with setup for each system in its own file. i18n for tests is now provided in the `jest.setup.js` file rather than imported in each test.

The way these default pages are being customized is by overwriting the "vocabularies" used by Amplify. This is also the way internationalization (i18n) is handled by Amplify. This initial PR is to set up the infrastructure of `authScreenLabels` plus the `en_us` localization of these labels, and call it in the `_app.js`.

Note: style and color customization is happening in another PR.

Out of scope: This PR does not attempt to move elements (e.g. "Forgot your password?" and "Or create an account"), or customize any pages outside the initial login.

Co-authored-by: Sawyer Hollenshead <git@sawyerh.com>