CP-146: Move design tokens directly into JS file (#82)
## 🎫 Ticket

[CP-146](https://lwd.atlassian.net/browse/CP-146)

## 🛠 Changes

This addresses an issue where the JS executes before the CSS has loaded, resulting in our `styleTokens()` method not being able to retrieve the design tokens being passed through the CSS (via the `html:after` pseudo selector).  

Not in love with this solution, but in the short term it feels like the least fragile. It could result in slight inconsistencies if we change a Sass variable and don't update it in the JS, but I'd rather that scenario than the theme not being applied at all.