Add MDX Storybook format for more freeform documentation (#125)
## 🛠 Changes

- Converted the `Title` Storybook file into [MDX format](https://github.com/storybookjs/storybook/blob/next/addons/docs/docs/mdx.md), as a test to see if this format would work for more freeform documentation, supporting additional Design documentation, like documenting design decisions & guidance.
- Renamed the script for running Storybook from `components` to `docs` to more accurately reflect the content in the Storybook.
- Updated a number of dependencies to better support `.mdx` format

## ℹ️ Context for reviewers

Up to this point, Storybook has been a tool used by the engineering team to document components for others on the engineering team. Its goal in this case is to answer questions like: 

- What components exist?
- What is this component for?
- What props does this component take?
- How does this component behave when rendered?

Engineers can also look at the source code to answer most of these questions, so its value for engineers is mostly to provide a more friendly UI to answer them.

There's also a world where Storybook can be used as a shared tool, supporting a more collaborative workflow between Engineering and Design teams, which I think is where a lot of its value begins to show. There's a desire from the Design team (and myself) to have a source of truth for design pattern documentation and guidance. Ideally that lives close to the actual implementation of the patterns. The goals for Storybook in this case would be to answer the questions above, as well as questions like:

- Why does this design pattern exist? What problems is it addressing? 
- What does this design pattern look like at different screen sizes?
- What have we learned about this pattern through user research? How has it evolved?
- What is the status of this pattern? Does it match what is currently in Figma?

These more extensive documentation pages would be a collaboration between Engineering and Design. Engineering's role would be close to what it is today: code the components, setup the basic Storybook page with an example of the component. It would then be Design's role to contribute design documentation, like any decisions or research that fed into the creation of the pattern, or guidance on different ways the component should be used. There may be times when Engineering and Design pair on a Storybook page to ensure the implementation achieves the intended design.