API-98: Mount local directory in development container (#120)
## 🎫 Ticket

[API-98](https://lwd.atlassian.net/browse/API-98)

## 🛠 Changes

- Add `dev` stage to Dockerfile to use for local development/development builds
- Mount project directory into container for local development with related config to run the development container as the host user to avoid permission issues

## ℹ️ Context for reviewers

Bind mounts (i.e., directly sharing a file between the host and the container) present a number of complications, probably the biggest being file permissions. On macOS, I'm given to understand Docker maps the root user in the container to the user running Docker, which avoids some of the permission issues. On Linux this is not the case. Can set up user namespaces to do this user mapping, but it's best practice to run as non-root anyway, so put some infrastructure in place to run as much as possible as non-root.

Following much of the approach laid out in this [Docker and Makefiles post](https://itnext.io/docker-makefile-x-ops-sharing-infra-as-code-parts-ea6fa0d22946).