API-117: Add a script to install Centrify AWS CLI (#141)
https://lwd.atlassian.net/browse/API-117

## 🛠 Changes

This adds a shell script to install the Centrify AWS CLI and a convenience wrapper command.

It works by:
- cloning the Centrify AWS CLI git repo into an `INSTALL_LOCATION`
- Downloading SSL certificates for centrify
- inserting a wrapper script to check python deps/work around some limitations, and 
- symlinking that wrapper to `/usr/local/bin` for convenience.

No, they don't seem to have an easy-install process. I know.

## ℹ️ Context for reviewers

Here's the expected workflow for developers that need terraform or AWS cli. This is also described in the infra README.

1. Run `./bin/install-centrify-aws-cli.sh <INSTALL_LOCATION>`
2. Run `login-aws`
3. Run `export AWS_PROFILE=AWS-498823821309-Infrastructure-Admin_profile` 
    or for James and I, `export AWS_PROFILE=AWS-498823821309-CloudOps-Engineer_profile`

To test that this works:
1. Install aws CLI if you don't have it: `pip install awscli`
2. Test aws CLI: `aws s3 ls`

Tested on Mac OSX and debian.