API-115: Add a script to bootstrap components for an environment (#157)
## 🎫 Ticket

https://lwd.atlassian.net/browse/API-115

## 🛠 Changes

This adds a script to bootstrap components for an environment. Each component is created separately.

```
$ ./bootstrap-env.sh test env-shared   # create non-app-specific environment components
$ ./bootstrap-env.sh test api                 # create API
$ ./bootstrap-env.sh test portal            # create Portal
```

I've also added `infra/pfml-aws/s3.tf` to standardize the config for environment S3 buckets (where tfstate files for that environment are stored). 

If a new environment is being created, add it to the list and the pfml-aws module can create it with the right configurations.

```tf
locals {
  environments = ["test", "staging"]
}

==> s3://pfml-test-env-mgmt/
==> s3://pfml-test-env-mgmt/
```

## ℹ️ Context for reviewers

This works by having templates for each component that are copied over to the `infra` directory.

```
bin/bootstrap-env/env-shared/<template files>
bin/bootstrap-env/api/<template files>
bin/bootstrap-env/portal/<template files>
```

These are almost word-for-word copied from the current sandbox environment, with some changes to the VPC lookups for `api`.

I'm not expecting these to be _totally_ up to date over time, but I'm hoping this fulfills the ticket's original intent of making it a little more straightforward for any given developer to know how to deploy the API to a new environment, for example.