CP-216: Add claim creation page to collect SSN (#168)
## 🎫 Ticket

https://lwd.atlassian.net/browse/CP-216

## 🛠 Changes

This adds a page to collect an employee's SSN.

## ℹ️ Context for reviewers

This uses the QuestionPage template to handle the boilerplate. It currently redirects from the date-of-birth page to this page, skipping the license page, in order to not orphan this page. Once the license page is checked in, the linking will be updated to intended routing.