API-113 (2/?): Add NLB and VPC Link for EOTSS VPCs (#172)
https://lwd.atlassian.net/browse/API-113

## 🛠 Changes

Adds an NLB and VPC Link for both the nonprod and prod EOTSS VPCs. This is like #171 but for the EOTSS account.

Also fixes the env-shared template which will refer to the NLB and VPC links defined in this PR.

## ℹ️ Context for reviewers

There's more info in the code-level documentation, but to clarify: an (NLB + VPC Link) acts as a private bridge between different API Gateways and a VPC. So:

- the nonprod pair will serve the (test, staging, prod-sandbox) API gateways, and
- the prod pair will serve the prod API Gateway.

## Demo

In env-shared:
Running the bootstrap script and running `terraform plan` after updating the nlb_port shows 13 resources to be created.

In aws:
Confirmed that the resources can be created.