CP-16: Add User-Agent Password Condition on s3 bucket policy (#152)
## 🎫 Ticket

[CP-16](https://lwd.atlassian.net/browse/CP-16)

## 🛠 Changes

- adds User-Agent header requirement to s3 bucket
- adds User-Agent header with random password to cloudfront headers

## ℹ️ Context for reviewers

Based on [this post](https://abridge2devnull.com/posts/2018/01/restricting-access-to-a-cloudfront-s3-website-origin/).

The goal of the PR is to only allow our cloudfront distribution to access the portal's s3 bucket. This PR originally set an Origin Access Identity on the distribution, but, as @doshitan, this would limit our ability to use normal website routing. 

We switched to setting a condition on our s3 bucket policy to only accept requests where the User-Agent stores a password shared between the bucket and the cloudfront distribution. 

## reviewers

@kyeah