API-135 (1/3): Generate RDS password on the fly (#210)
https://lwd.atlassian.net/browse/API-135

Fully manage RDS password through terraform by generating it on the fly.

Since we're grabbing and decoding the password from SSM currently, the password is already being stored in tfstate. This change does not provide any security benefits, but reduces the manual steps needed for setting up a new database instance.

This will likely require the sandbox database to be recreated, which was likely anyways as we finetune the configuration.