CP-300: Support entry of leading zeros for month and day fields (#243)
## 🎫 Ticket

https://lwd.atlassian.net/browse/cp-300

## 🛠 Changes

- Updates the `onChange` handler of `InputDate` so that it doesn't add leading zeros to the `value` so that a user can type "1" or "01"
- Adds an `onBlur` handler to the month/day/year fields so that the date can be stored in our form state with a valid ISO 8601 date string, with leading zeros.

## ℹ️ Context for reviewers

This was more complicated than I anticipated but I had trouble thinking of a different solution. We're working with the following requirements:

- The API requires date values to be valid ISO 8601 date strings (`YYYY-MM-DD`), so this is the format we need to store in our form state.
- We want to allow a user to enter a single digit month / day value (1–9)
- We want to allow a user to enter a single digit month / day value with a leading zero (01–09)
- When a user returns to date field with an existing value, it will always have a leading zero because the full date is stored as ISO 8601.

State flows like this in our App:

```
├── <Page component>
           └── formState
           └── <InputDate value={formState.date} />
                      └──  values = parseDateParts(props.value)
                      └──  <InputText name="month" value={values.month}
                      └──  <InputText name="day" value={values.day}
                      └──  <InputText name="year" value={values.year}
```

Before this change, we could call `onChange` with a valid ISO 8601 string with leading zeros, and our `parseDateParts` would remove leading zeros. Updating `parseDateParts` so it no longer removes a leading zero means that, without any other changes, when a user enters "1", the field will change to "01". That would be an issue because it would mean they wouldn't be able to type "12". (In #241, we are introducing a change that prevents the user from entering more than 2 digits for months and days.)

This presented a dilemma: how do we still store valid ISO 8601 date strings in our state, while allow someone to enter digits like 1 and also enter digits greater than 9 😅. The solution in this PR is to not add leading zeros to the value we store in our state until the user is no longer actively editing a date field, as identified by `onBlur`