API: Skip broken /users tests (#257)
In a recent PR[1] the handlers for `/users` routes were changed to use the
database (yay), but without support for mocking the database (oh no) and without
cleaning up after test runs (eek). So user tests write real user records to the
real database that stick around after the tests are complete.

Not only is this bad practice, but it also breaks the tests. Until we put the
foundations in place to run database tests correctly, disable them.

Also, some of the tests were updated to be incorrect, such as testing requests
with missing required parameters, which should be a 400 response, not a 500.

[1] https://github.com/EOLWD/pfml/pull/205