CP-251: Add In Progress dashboard state (#272)
## 🎫 Ticket

https://lwd.atlassian.net/browse/cp-251

## 🛠 Changes

- When a claim exists in the local app state, display the In Progress dashboard state
- Created "Screens -> Dashboard" Storybook page for previewing the different dashboard states
- Add `className` prop to `ButtonLink` component

## ℹ️ Context for reviewers

The in progress dashboard can't yet be previewed in the app, but you can preview it in Storybook. Once [CP-294](https://lwd.atlassian.net/browse/CP-294) is complete, the in progress dashboard should work.

This is a quick implementation to support the pilot. The dashboard is likely to change a decent bit so I went with the quickest implementation for now.