CP-346: Map all account creation password validation errors to single message key (#298)
## 🎫 Ticket

https://lwd.atlassian.net/browse/CP-346

## 🛠 Changes

Adds a collection of account creation password validation errors to the `customAmplifyErrorMessageKeyMap` function, and maps all those errors to a single message key.

## ℹ️ Context for reviewers

Per the details in the ticket, Cognito does not return all the relevant error messages, but only one or two. This can result in a whack-a-mole experience where if you only solve the error shown, you could still have another error in the next round.

This goes in hand with @sawyerh 's work to add hint text to the account creation password field that details the password requirements.