Add useCollectionState hook and add static methods to Collection class (#300)
## 🎫 Ticket

https://lwd.atlassian.net/browse/cp-294

## 🛠 Changes

- Adds useCollectionState hook
- Adds static methods to Collection that allow collections to be immutable 

## ℹ️ Context for reviewers

This PR is split from #284 to more easily review. Please see discussion there.

Rationale behind moving to useCollection:

The Collection silently mutated the application's state with out triggering React component lifecycle events.

With the setState methods provided by the useState hook, React re-renders the component whenever the method is called. This keeps the views in sync with the component's state. Items are added to a Collection, on the other hand, by calling collection.add(item) which modifies the collection's internal byId property. The React component has no way of knowing the state has changed, and DOM elements can get out of sync with the component's state

## Reviewers
@sawyerh