CP-335: Add request module and currentUser method (#308)
## 🎫 Ticket

https://lwd.atlassian.net/browse/cp-335

## 🛠 Changes

- Added `request` module for performing API requests, which is mostly a wrapper around [`fetch`](https://developer.mozilla.org/en-US/docs/Web/API/Fetch_API/Using_Fetch)
- Added `currentUser` method to the `usersApi` module. ([Swagger doc](https://67385ye4yb.execute-api.us-east-1.amazonaws.com/test/api/docs/#/Users/massgov.pfml.api.users.users_current_get))

## ℹ️ Context for reviewers

This is a first of a few PRs. In this PR, I'm primarily looking for feedback on what folks think about this `request` module approach. I would prefer not using an external request library since those tend to be heavier than what we need, and I believe we can get what we need from using just the platform. Also, security 😄