CP-352 Clear Amplify errors when user starts the reset password flow (#332)
## 🎫 Ticket

https://lwd.atlassian.net/browse/CP-352

## 🛠 Changes

* When a user begins the reset password flow we clear any existing Amplify errors. This is done using Amplify's Hub utility

## ℹ️ Context for reviewers

The CustomForgotPassword component we use in Amplify involves two screens but Amplify treats them as a single authState change. So a user has to complete both screens of the CustomForgotPassword component before it triggers an authState change in the Authenticator component.

Because authState changes are currently how we reset Amplify errors this creates an issue when a user triggers an error alert on the first screen of the forgot password flow (that is -- when they're asked for an email address for which to reset the password). If a user triggers an error on the first screen and then successfully moves to the second screen the error remains.

You can see where the `forgotPassword` even is triggered in Amplify [here](https://github.com/aws-amplify/amplify-js/blob/14f1ef9b88bc5cb99ec0a47308d3fb9cf4c89102/packages/auth/src/Auth.ts#L1571).

### Alternative considered

I also considered just using Hub listening as the mechanism to reset Amplify errors. Unfortunately certain authState changes don't seem to include a Hub event (eg clicking the "Return to login" button). It seems like we really do need to handle all authState changes as well as this particular Hub event.

## 👀 Demo

![cp-352-demo](https://user-images.githubusercontent.com/9039672/81101490-3d436e00-8edc-11ea-97a1-485a29dfeafe.gif)