CP-201: Create initial page for attestation of employer notification (#339)
## 🎫 Ticket

https://lwd.atlassian.net/browse/CP-201

## 🛠 Changes

This adds a page for users to attest to having notified their employer. It also updates the Duration page to point to this page, and moves to shared messages for "yes" and "no" choice fields.

## ℹ️ Context for reviewers

This does not connect to the API. That will be added in a subsequent PR, along with aligning field names with the API format.

I am using "Medical Leave Application" as the placeholder title for this page per approval from @hanya-m . As this is an initial pass at the page, that title may change in a final version.