API-151: Run migrations during API deployment (#333)
https://lwd.atlassian.net/browse/API-151

Adds a migration step to the API deployment workflow.

This implementation runs migrations after the terraform apply / ECS deployment. This was the easiest way for now — the ECS task is managed within the API terraform, and we depend on the terraform output variables too.

I could see us wanting to optimize for the opposite case in the future (running migrations first), but things are pretty unstable at the moment so it seems unlikely to be much difference in the near term.

`aws ecs wait tasks-stopped` does not error out if the task fails. But let's leave that for another PR.