API: Use Flask's request context for access to database (#297)
Global things always feel like a code smell.

For the route handlers use Flask's context stuff to pass dependencies to it as there seems to be few alternatives. Other things (e.g., service layers, utils, etc.) would then accept a DB session as a parameter, create one as needed, or also require running in an application context (if this is how we want to handle dependency injection in general).