Use shab for bootstrap-env templates (#418)
shab[1] is a minimal bash-only templating language that basically applies bash
variable substitutions/expansions in the template. That does mean the template
content has to be shell escaped, for Terraform that generally means the `$` used
for string interpolation. A mild annoyance, but also means we don't have to do a
bunch of individual `sed` statements nor have to figure out a cross-platform
in-place `sed` command.

[1] https://github.com/zimbatm/shab