CP-388 Add FileCardList component (#415)
## 🎫 Ticket

https://lwd.atlassian.net/browse/CP-388

## 🛠 Changes

* Adds a new FileCardList component
* Refactors the FileCard a bit so that it retrieves the filename from the File object, not a separate `filename` prop

## ℹ️ Context for reviewers

Uses the FileCard component added in CP-387. The Figma design is [here](https://www.figma.com/file/v8LlmK8r1JmByqtNVMvqjS/PFML?node-id=2566%3A4376).

The FileCards are rendered in an unstyled list so that it's easier to understand using a screen reader. The input HTTP element is rendered as a transparent input element shifted offscreen for the same reason (it still appears as an interact element to a screen reader).

The File constructor isn't available in Internet Explorer which breaks the example FileCards loaded as part of the component stories. I've added an error message for when that (or other errors) happen and telling the user to just select file(s) themselves to see the FileCards.

I did some light testing in Chrome and Safari on osx, Internet Explorer on Windows, and Chrome on Android and other than the File constructor issue it worked as expected.

## 👀 Demo

![cp-388-storybook](https://user-images.githubusercontent.com/9039672/82917179-127b9100-9f41-11ea-9cf0-7843b8accce9.gif)