CP-291: Setup deploy pipeline for Storybook (#437)
🎫 Ticket
https://lwd.atlassian.net/browse/cp-291
https://lwd.atlassian.net/browse/cp-292

🛠 Changes
Added S3 bucket in the Test environment for hosting the Storybook site
Added GitHub action for deploying the Storybook site when changes are merged to master

ℹ️ Context for reviewers

I don't currently have a use case for needing a Storybook site for every environment, so constrained this to just hosting in the Test environment. We can pretty easily expand this if a use case appears.

I needed a mindless task to work on and this was mostly copying + modifying existing code. Storybook has also come up in some recent convos, and will be useful for sharing with stakeholders and Mayflower.

👀 Demo
Site will exist here once merged: http://massgov-pfml-test-storybook-builds.s3-website-us-east-1.amazonaws.com/