API: Fix id column sequence on seeded lookup tables (#400)
The id columns are set to use an autoincremeting integer, but this sequence is
not updated if the column's value is set explicitly. The seeds explicitly set
the column value, but do not update the sequence for where they leave off,
causing an error when a lookup value is created by the application (e.g., the
user status stuff via `get_or_make_status`). So fix that for now.

We have to figure out a consistent way to manage the lookup values going
forward, because any changes to `lk_status` will break user creation atm.