API-114: Link SSL certificates to Portal and API (#433)
It turns out that the Cert manager API behaves differently from when you do it in the console. In the console, it does the proper MX lookups to find real mass.gov validation emails, so we have to do it manually through the console.

This removes the code added in #351 and does a lookup for the manually-created certs, which are then linked to the Portal and API Gateway applications.