API-123: Create User record when users confirm sign up via Cognito (#410)
Create Lambda to connect to Cognito's Post Confirmation trigger to create User
record in API database automatically on sign up confirmation.

The Terraform for the Lambda and Cognito are decoupled, so that API and Portal
infrastructure don't have a hard dependency on each other. This does require
some variables to be coordinated manually as we have been doing with
`cors_origins`. On the API side, the Cognito Pool ID for the environment needs
set to finalize permission for the Lambda (so only the specified pool can use
it). On the Portal side, the ARN of the Lambda needs set to actually hook it up
to Cognito (i.e., the Portal infrastructure controls if the Post Confirmation
hook is enabled or not).

Since all the Cognito User Pools already exist, we can set those variables on
the API side. Enabling the hook should be done separately and more
incrementally, since the API infrastructure needs to deploy first, to create the
Lambda, then the Portal infrastructure needs to deploy to connect the Lambda to
Cognito.

https://lwd.atlassian.net/browse/API-123