CP-363: Connect confirm & submit page to API module (#450)
## 🎫 Ticket

https://lwd.atlassian.net/browse/CP-363

## 🛠 Changes

This adds a mocked method to the claimsApi: `submitClaim`. It also connects the `handleSubmit` method to the claimsApi, and wraps the `confirm` page with the `withClaim` hook. It adds testing for both.

It also routes the `notified-employer` page to this final `confirm` page.

## ℹ️ Context for reviewers

This is a follow-on for https://github.com/EOLWD/pfml/pull/420, which added all the visual front-end code.

## 👀 Demo

No visual changes.