Set default for unrestrictedClaimFlow to true (#455)
## 🎫 Ticket

No ticket. At request of Nicole Budzius, PM for Claimant Portal.

## 🛠 Changes

This sets the flag `unrestrictedClaimFlow` to true in all environments. This will unlock the full claim flow in staging and prod.

## ℹ️ Context for reviewers

This is the first of three steps:
*1) Flip flag to true in all environments*
2) Remove use of flag in code
3) Delete flag from config entirely.

Each step can proceed after the previous step has been fully deployed to production.