Delete unrestrictedClaimFlow flag (#457)
## 🎫 Ticket

No ticket. At request of Nicole Budzius, PM for Claimant Portal.

## 🛠 Changes

This removes the `unrestrictedClaimFlow` flag from the list of feature flags, as it is no longer in use.

## ℹ️ Context for reviewers

This is the third of three steps:
1. Flip flag to true in all environments (deployed to prod)
2. Remove use of flag in code (deployed to prod)
3. Delete flag from config entirely ◀️ 

Each step can proceed after the previous step has been fully deployed to production.