Enable $utilities-use-important (#453)
🛠 Changes
Configure the U.S. Web Design System so that its utility classes always override a component's classes, regardless of the component selector specificity.

ℹ️ Context for reviewers
This setting results in utility style declarations to receive the !important rule, which addresses an annoyance I've experienced a number of times: adding a utility class to a component sometimes doesn't apply the styling/spacing you expect it to, because the component's class has a higher specificity than the utility class. With the !important rule, now utility classes will always override a style like you'd expect it to.