API-239 Add Bouncer for Authorization and Create Related Modules (#502)
https://lwd.atlassian.net/browse/API-239

🛠 Changes
added flask-bouncer to the app.
created rules.py: a module for defining all the rules for each role
created flask.py: a wrapper module for flask-bouncer
added an example for setting up authorization with the users/GET route (In comments)
ℹ️ Context for reviewers
This PR is for setting up the groundwork for authorization in pfml. flask-bouncer has the advantage of allowing us to define all of the rules regarding each user (e.g. user, employee, employer, etc) in one place. Initial set up is in app.py, this section wraps bouncer around the app and marks define_authorization as the method that defines user permissions. An important note from the docs:

flask-bouncer by default looks for current_user or user stored in flask’s g

We will need to use the @bouncer.user_loader decorator if we want to override this behavior.