API-255 (2/2): Remove lk_status table (#491)
https://lwd.atlassian.net/browse/APi-255

This removes the status lookup table (which was being shared across tables). The following tables are affected:

- continuous_leave_period: replaced with is_estimated, default true.
- reduced_schedule_leave_period: replaced with is_estimated, default true.
- application: No replacement added. We already have started_time, completed_time, and submitted_time, which are sufficient for deriving the state.

ℹ️ Context for reviewers

Based on the conversation in https://lwd.atlassian.net/browse/API-288 with the Portal team, we will derive a status field in the ApplicationResponse model. For now, we will keep the OpenAPI schema the way it already is (no status in the schema.)