API: Bump local DB connection limit to fix test errors (#565)
This helps avoid test errors of:

```
sqlalchemy.exc.OperationalError: (psycopg2.OperationalError) FATAL:  sorry, too many clients already
```

We create and destroy DB connections a bunch when running the tests. There may
be something we can do on the app side to signal to the server to be more
aggressive about pruning clients, but this is a simple fix.

Also don't rely on the PostgreSQL Docker container's special feature of passing
things that look like flags to the `postgres` command, just fully specify the
command we want to run. Saves having to write a comment explaining it.