API-303: Fix updating currently authenticated user (#547)
Due to our current SQLAlchemy settings of `expire_on_commit=False`, models that
have already been fetched into the session do not get marked to update from the
database automatically when commits happen.

So for cases where updates are not made through the model object itself,
the in-memory copy will fall out of date with the database.

This has manifested itself in the `PATCH /users/{user_id}` route when updating
the authenticated user.

On successful authentication for a request, the app fetches the corresponding
User model and stores it on `g.current_user`, so that User model is now part of
the current db session.

The `PATCH` request handler would directly issue an `update()` query based on
the request body, which would successfully update values in the database, but
not update any existing model in the db session.

To then retrieve the updated User model to return in the response, a `get()`
would be issued for the updated user id. But `get()`[1] is special in
SQLAlchemy, if it already has the non-expired model corresponding to the
requested id, it won't check the database, just return the in-memory copy,
quoting from docs:

> Query.get() is special in that it provides direct access to the identity map
> of the owning Session. If the given primary key identifier is present in the
> local identity map, the object is returned directly from this collection and
> no SQL is emitted, unless the object has been marked fully expired. If not
> present, a SELECT is performed in order to locate the object.

So for the case where a `PATCH` is being processed for the currently
authenticated user, whose User model is already loaded into the db session, the
values in the database would be updated, but `get()` would return the version in
memory, which had the previous values.

There are a number of ways to fix this particular bug. Updating the `PATCH`
request handler to follow a general pattern of first fetching the model to be
updated, updating the model itself, and persisting the changes that way would
seem to be the clearest and easily lends itself to only processing fields that
have been explicitly set in the body, which is another property we want `PATCH`
endpoints to have.

For this particular request the latter part is not super critical as there is
only one field that can be updated, but that may change overtime and also this
is a general pattern we should follow for `PATCH` endpoints, so is more
consistent.

As the app evolves we may need to revisit the `expire_on_commit` setting or
provide better guidance/tooling to manually refresh/expire models that could be
touched by updates that did not go through the model itself.

[1] https://docs.sqlalchemy.org/en/13/orm/query.html#sqlalchemy.orm.query.Query.get

https://lwd.atlassian.net/browse/API-303